@@ -8,13 +8,13 @@
     "@testing-library/user-event": "^12.8.3",
     "@types/jest": "^26.0.21",
     "@types/node": "^12.20.6",
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "@types/react": "^19.0.1",
+    "@types/react-dom": "^19.0.2",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-router": "^7.2.0",
+    "react-router-dom": "^7.2.0",
+    "react-scripts": "5.0.1",
     "typescript": "^4.2.3",
     "web-vitals": "^1.1.1"
   },
@@ -44,6 +44,6 @@
   },
   "devDependencies": {
     "@types/react-router-dom": "^5.1.7",
-    "typescript-plugin-css-modules": "^3.2.0"
+    "typescript-plugin-css-modules": "^5.1.0"
   }
 }
